pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "open-memory-capsule"
version = "0.1.0"
description = "Capture everything. Remember forever. Search naturally."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["memory", "capture", "search", "ai", "rag", "whatsapp", "telegram"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "chromadb>=0.4.24",
    "openai-whisper>=20231117",
    "torch>=2.1.0",
    "easyocr>=1.7.1",
    "pymupdf>=1.23.0",
    "Pillow>=10.0.0",
    "watchdog>=4.0.0",
    "telethon>=1.34.0",
]

[project.optional-dependencies]
openai = ["openai>=1.12.0"]
anthropic = ["anthropic>=0.20.0"]
groq = ["groq>=0.5.0"]
slack = ["slack-bolt>=1.18.0"]
discord = ["discord.py>=2.3.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]
all = [
    "openai>=1.12.0",
    "anthropic>=0.20.0",
    "groq>=0.5.0",
    "slack-bolt>=1.18.0",
    "discord.py>=2.3.0",
]

[project.scripts]
capsule = "cli.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/open-memory-capsule"
Issues = "https://github.com/yourusername/open-memory-capsule/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
